\name{JM4QTN-package}
\alias{JM4QTN-package}
\alias{JM4QTN}
\docType{package}
\title{
Joint Mapping for QTN
}
\description{
The "JM4QTN" package is based on composite interval mapping and regression analysis to do association analysis and linkage analysis for multiple-line cross population and multiple traits. The package can also identify pleiotropic or linked QTL.
}
\details{
\tabular{ll}{
Package: \tab JM4QTN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-23\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
JunhuLi,GuoliangLi,HeliChen,KunCheng,WenxinLiu

Maintainer: JunhuiLi<junhuili@cau.edu.cn>
}
\references{
Junhui Li, Haixiao Hu, Yujie Meng, Kun Cheng, Guoliang Li, Wenxin Liu, and Shaojiang Chen.(2016)Pleiotropic QTL detection for stalk traits in maize and related R package programming. Journal of China Agricultural University. DOI 10.11841/j.issn.1007-4333.2016.06.00(in chinese)
}
\keyword{ package JointMapping QTN}